Add UIActionUtilities, move EnumerateUIActionHandlers and TryExecute into this class.
