+++ Sandra.UI/UIActionForm.cs
@@ -21,7 +21,6 @@
 
 using Eutherion.Win;
 using Eutherion.Win.UIActions;
-using Eutherion.Win.Utils;
 using System;
 using System.Windows.Forms;
 
@@ -42,7 +41,7 @@ namespace Sandra.UI
             try
             {
                 // This code makes shortcuts work for all UIActionHandlers.
-                return KeyUtilities.TryExecute(keyData, FocusHelper.GetFocusedControl())
+                return UIActionUtilities.TryExecute(keyData, FocusHelper.GetFocusedControl())
                     || base.ProcessCmdKey(ref msg, keyData);
             }
             catch (Exception e)
